refactor: centralize date navigation and fix today button timezone handling
