+++ SparkyFitnessFrontend/src/pages/Diary/Diary.tsx
@@ -1,17 +1,9 @@
 import { useState, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
-import { Button } from '@/components/ui/button';
-import { Calendar } from '@/components/ui/calendar';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
-import { CalendarIcon, ChevronLeft, ChevronRight } from 'lucide-react';
-import { cn, formatDateToYYYYMMDD } from '@/lib/utils';
 import { useActiveUser } from '@/contexts/ActiveUserContext';
 import { usePreferences } from '@/contexts/PreferencesContext';
+import DayNavigator from '@/components/DayNavigator';
 import DiaryTopControls, { DayTotals } from './DiaryTopControls';
 import MealCard from './MealCard';
 import ExerciseCard from './ExerciseCard';
@@ -45,31 +37,23 @@ import {
   useFoodEntries,
   useFoodEntryMeals,
 } from '@/hooks/Diary/useFoodEntries';
+import { todayInZone } from '@workspace/shared';
 
 const Diary = () => {
   const { t } = useTranslation();
   const { activeUserId } = useActiveUser();
   const location = useLocation();
   const navigate = useNavigate();
-  const {
-    formatDate,
-    formatDateInUserTimezone,
-    getDateRelationToToday,
-    parseDateInUserTimezone,
-    loggingLevel,
-    energyUnit,
-    convertEnergy,
-  } = usePreferences();
+  const { timezone, loggingLevel, energyUnit, convertEnergy } =
+    usePreferences();
   const [editingEntry, setEditingEntry] = useState<FoodEntry | null>(null);
   const [editingFoodEntryMeal, setEditingFoodEntryMeal] =
     useState<FoodEntryMeal | null>(null); // State for editing logged meal entry
   const [searchParams, setSearchParams] = useSearchParams();
 
   const [selectedDate, setSelectedDate] = useState(
-    searchParams.get('date') ??
-      formatDateInUserTimezone(new Date(), 'yyyy-MM-dd')
+    searchParams.get('date') ?? todayInZone(timezone)
   );
-  const [date, setDate] = useState(parseDateInUserTimezone(selectedDate));
   debug(loggingLevel, 'FoodDiary component rendered for date:', selectedDate);
   const [exercisesToLogFromPreset, setExercisesToLogFromPreset] = useState<
     PresetExercise[] | undefined
@@ -121,7 +105,6 @@ const Diary = () => {
     : [];
 
   const dayTotals = calculateDayTotals(foodEntries, foodEntryMeals);
-  const selectedDateRelation = getDateRelationToToday(selectedDate);
 
   // Handle navigation for opening food search dialog
   useEffect(() => {
@@ -185,31 +168,6 @@ const Diary = () => {
     } finally {
       setIsCopyDialogOpen(false);
     }
-  };
-
-  const handleDateSelect = (newDate: Date | undefined) => {
-    debug(loggingLevel, 'Handling date select:', newDate);
-    if (newDate) {
-      setDate(newDate);
-      const dateString = formatDateToYYYYMMDD(newDate);
-      info(loggingLevel, 'Date selected:', dateString);
-      setSelectedDate(dateString);
-      setSearchParams({ date: dateString });
-    }
-  };
-
-  const handlePreviousDay = () => {
-    debug(loggingLevel, 'Handling previous day button click.');
-    const previousDay = new Date(date);
-    previousDay.setDate(previousDay.getDate() - 1);
-    handleDateSelect(previousDay);
-  };
-
-  const handleNextDay = () => {
-    debug(loggingLevel, 'Handling next day button click.');
-    const nextDay = new Date(date);
-    nextDay.setDate(nextDay.getDate() + 1);
-    handleDateSelect(nextDay);
   };
 
   const handleFoodSelect = async (item: Food | MealType, mealType: string) => {
@@ -330,90 +288,13 @@ const Diary = () => {
   if (loading) return <div>Loading...</div>;
   return (
     <div className="space-y-6">
-      {/* Date Navigation */}
-      <div className="grid grid-cols-[minmax(0,1fr)_auto_minmax(0,1fr)] items-center mb-5 gap-2">
-        <div className="flex justify-end">
-          {selectedDateRelation !== 'today' && (
-            <Button
-              variant="ghost"
-              size="sm"
-              className="text-xs text-muted-foreground h-9 px-3 rounded-full border border-border/60"
-              onClick={() => handleDateSelect(new Date())}
-            >
-              Today
-            </Button>
-          )}
-        </div>
-        <div
-          className={cn(
-            'relative flex items-center gap-0 rounded-full border border-border/60 bg-background overflow-hidden transition-colors',
-            selectedDateRelation === 'past' && 'border-date-past/50',
-            selectedDateRelation === 'future' && 'border-date-future/50'
-          )}
-        >
-          {selectedDateRelation !== 'today' && (
-            <div
-              className={cn(
-                'absolute inset-0 pointer-events-none',
-                selectedDateRelation === 'past' && 'bg-date-past/10',
-                selectedDateRelation === 'future' && 'bg-date-future/10'
-              )}
-            />
-          )}
-          <Button
-            variant="ghost"
-            size="icon"
-            onClick={handlePreviousDay}
-            className="relative h-9 w-9 rounded-none border-r border-border/60"
-          >
-            <ChevronLeft className="h-4 w-4" />
-          </Button>
-          <Popover>
-            <PopoverTrigger asChild>
-              <Button
-                variant="ghost"
-                className="relative h-9 px-4 rounded-none font-normal text-sm gap-2"
-              >
-                <CalendarIcon
-                  className={cn(
-                    'h-3.5 w-3.5',
-                    selectedDateRelation === 'past' && 'text-date-past',
-                    selectedDateRelation === 'future' && 'text-date-future'
-                  )}
-                />
-                {date ? (
-                  formatDate(date)
-                ) : (
-                  <span className="text-muted-foreground">
-                    {t('foodDiary.pickADate', 'Pick a Date')}
-                  </span>
-                )}
-              </Button>
-            </PopoverTrigger>
-            <PopoverContent
-              className="w-auto p-0"
-              align="center"
-              sideOffset={8}
-            >
-              <Calendar
-                mode="single"
-                selected={date}
-                onSelect={handleDateSelect}
-                yearsRange={10}
-              />
-            </PopoverContent>
-          </Popover>
-          <Button
-            variant="ghost"
-            size="icon"
-            onClick={handleNextDay}
-            className="relative h-9 w-9 rounded-none border-l border-border/60"
-          >
-            <ChevronRight className="h-4 w-4" />
-          </Button>
-        </div>
-        <div />
-      </div>
+      <DayNavigator
+        selectedDate={selectedDate}
+        onDateChange={(dateString) => {
+          setSelectedDate(dateString);
+          setSearchParams({ date: dateString });
+        }}
+      />
 
       {/* Top Controls Section */}
       {goals && (
